8293254: x86_64: Extract arraycopy stub implementations into a separate file
