+++ src/hotspot/cpu/x86/macroAssembler_x86.hpp
@@ -2047,42 +2047,14 @@ public:
   void cache_wb(Address line);
   void cache_wbsync(bool is_pre);
 
-#if COMPILER2_OR_JVMCI
-  void arraycopy_avx3_special_cases(XMMRegister xmm, KRegister mask, Register from,
-                                    Register to, Register count, int shift,
-                                    Register index, Register temp,
-                                    bool use64byteVector, Label& L_entry, Label& L_exit);
-
-  void arraycopy_avx3_special_cases_conjoint(XMMRegister xmm, KRegister mask, Register from,
-                                             Register to, Register start_index, Register end_index,
-                                             Register count, int shift, Register temp,
-                                             bool use64byteVector, Label& L_entry, Label& L_exit);
-
-  void copy64_masked_avx(Register dst, Register src, XMMRegister xmm,
-                         KRegister mask, Register length, Register index,
-                         Register temp, int shift = Address::times_1, int offset = 0,
-                         bool use64byteVector = false);
-
-  void copy32_masked_avx(Register dst, Register src, XMMRegister xmm,
-                         KRegister mask, Register length, Register index,
-                         Register temp, int shift = Address::times_1, int offset = 0);
-
-  void copy32_avx(Register dst, Register src, Register index, XMMRegister xmm,
-                  int shift = Address::times_1, int offset = 0);
-
-  void copy64_avx(Register dst, Register src, Register index, XMMRegister xmm,
-                  bool conjoint, int shift = Address::times_1, int offset = 0,
-                  bool use64byteVector = false);
-
+#ifdef COMPILER2_OR_JVMCI
   void generate_fill_avx3(BasicType type, Register to, Register value,
                           Register count, Register rtmp, XMMRegister xtmp);
-
 #endif // COMPILER2_OR_JVMCI
 
   OopMap* continuation_enter_setup(int& stack_slots);
   void fill_continuation_entry(Register reg_cont_obj, Register reg_flags);
   void continuation_enter_cleanup();
-
 #endif // _LP64
 
   void vallones(XMMRegister dst, int vector_len);
